20210701_Update
- Gauge Data Connecting 구현

  : 데이터를 받온 데이터를 1씩 증가시키는 방향으로 구현

  : 바로바로 대입할 시에 부자연스럽게 뚝뚝 끊김

  : QTimer의 주기를 변경해봤는데 이거는 하드웨어 연결을 통해 조율이 필요할거같음

    (100ms -> 50ms)

  : 주기 조율이 어려울 경우에는 데이터 증감률을 늘려야할것같음. 현재는 1

- Arduino Demo File 작업

  : Demo Test를 위한 코드 구현

  : 아직 보드 테스트 미완료. Endian에 따라 Code 수정이 필요할 수도 있음

+++ monitor_Packet.c
@@ -1,6 +1,6 @@
 #include "monitor_Packet.h"
 
-packet Encoding_Packet(uint8_t rw, uint8_t id, uint8_t data){
+packet Encoding_Packet(uint8_t rw, uint8_t id, uint16_t data){
     packet returnPacket;
     returnPacket.R.rw = rw;
     returnPacket.R.id = id;
